pyproject.toml
[project]
name = "gb-price-crwl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "gspread>=6.2.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "seleniumbase>=4.38.0",
]
